a36ab106bba9728afc9976db3d194d6ea7a5c414 fix(types): use f64 for GetMiningInfo networkhashps field (deadmanoz)

Pull request description:

  As per raised in https://github.com/rust-bitcoin/corepc/issues/429, addressing type discrepancies between corepc-types and Bitcoin Core.

  Bitcoin Core has always returned `networkhashps` as a double since `getmininginfo` was introduced. The previous type (`i64`) causes deserialisation to fail. Here's a minimal example:

  ```rust
  use corepc_client::client_sync::v28::Client;
  use corepc_client::client_sync::Auth;

  fn main() -> Result<(), Box<dyn std::error::Error>> {
      let auth = Auth::UserPass("bitcoin".to_string(), "bitcoin".to_string());
      let client = Client::new_with_auth("http://127.0.0.1:8332", auth)?;
      let info = client.get_mining_info()?;
      println!("networkhashps: {}", info.network_hash_ps);
      Ok(())
  }
  ```

  ```
  Error: JsonRpc(Json(Error("invalid type: floating point `1.055282156351532e+21`, expected i64", line: 1, column: 179)))
  ```

  This change:
  - Changes the `networkhashps` field type from `i64` to `f64`
  - Modifies existing integration test to verify correct deserialisation (mines some blocks to yield a `networkhashps` `> 0.0`) and remove redundant comments

ACKs for top commit:
  tcharding:
    ACK a36ab106bba9728afc9976db3d194d6ea7a5c414

Tree-SHA512: 26436d4e76e421920442e8cc201cdcbaeaa59d17d6906218c23737971c1e91bb61970c40bd69ce1c65215801d2d9ec74df8a6c97afdbaefe63437336b4cab6a7

+++ types/src/v30/mining/mod.rs
@@ -38,7 +38,7 @@ pub struct GetMiningInfo {
     pub target: String,
     /// The network hashes per second.
     #[serde(rename = "networkhashps")]
-    pub network_hash_ps: i64,
+    pub network_hash_ps: f64,
     /// The size of the mempool.
     #[serde(rename = "pooledtx")]
     pub pooled_tx: i64,
